git-svn-id: http://svn.mkgmap.org.uk/splitter/trunk@197 838fabe5-b1a8-4b81-9896-7a84386980e9

+++ src/uk/me/parabola/splitter/BinaryMapWriter.java
@@ -255,7 +255,7 @@ public class BinaryMapWriter extends OSMWriter {
           Osmformat.Way.Builder bi = Osmformat.Way.newBuilder();
           bi.setId(i.getId());
           long lastid = 0;
-          for(int j : i.getRefs()) {
+          for(long j : i.getRefs()) {
             long id = j;
             bi.addRefs(id - lastid);
             lastid = id;
